edit type conversion, gcc  ok!

+++ ft_substr.c
@@ -15,13 +15,13 @@
 char	*ft_substr(char const *s, unsigned int start, size_t len)
 {
 	char	*substr;
-	int		i;
+	unsigned int	i;
 
 	substr = (char *)malloc(sizeof(char) * (len + 1));
 	if (!substr)
 		return (NULL);
 	i = 0;
-	while (i < len && start < ft_strlen((char *)s) && s[start + i] != '\0')
+	while (i < len && start < (unsigned int)ft_strlen((char *)s) && s[start + i] != '\0')
 	{
 		substr[i] = s[start + i];
 		i++;
